==> scripts/01_preprocess.ipynb <==
## MDS Thesis
#### 01. Pre-process the PARTYPRESS data

<br>
<hr style="opacity: 0.5">

### Setup

# load libraries
import os
import pandas as pd
import pickle
import nltk
nltk.download('punkt')

from datasets import Dataset
from nltk.tokenize import sent_tokenize

# set wd
os.getcwd()
ROOT_DIR = "/Users/varvarailyina/hertie/mds_thesis/data"

# load labels data
df_partypress = pd.read_csv(os.path.join(ROOT_DIR, "in", "partypress", "csv", "partypress.csv"))

# load text data
df_texts = pd.read_csv(os.path.join(ROOT_DIR, "in", "partypress", "csv", "partypress_texts.csv"))

<hr style="opacity: 0.5">

### Pre-process data

-- *Merge PARTYPRESS datasets*

# check indices
print(df_partypress.index.is_unique)
print(df_texts.index.is_unique)

# set merging index to be 'id'
df_partypress.set_index('id', inplace=True)
df_texts.set_index('id', inplace=True)

# merge datasets based on id and country_name
df = df_partypress.merge(df_texts, on=['id', 'country_name'], how='inner')

# check var names
print(df.columns)

-- *Filter data*

# filter for germany
df_de = df[df['country_name'] == 'germany']

# filter out press releases with no issue (use Monolingual Transformer)
df_de = df_de[df_de['issue_mono'].isin([98, 99]) == False]

df_de.info()

-- *Add issue lables*

# specify issue labels
issue_mapping = {
    1: "Macroeconomics",
    2: "Civil Rights",
    3: "Health",
    4: "Agriculture",
    5: "Labor",
    6: "Education",
    7: "Environment",
    8: "Energy",
    9: "Immigration",
    10: "Transportation",
    12: "Law and Crime",
    13: "Social Welfare",
    14: "Housing",
    15: "Domestic Commerce",
    16: "Defense",
    17: "Technology",
    18: "Foreign Trade",
    20: "Government Operations",
    23: "Culture",
    191: "International Affairs",
    192: "European Integration"
}

# add `issue_label` column
df_de['issue_label'] = df_de['issue_mono'].map(issue_mapping)

# define needed variables
needed_vars = [
    'country_name', 'party', 'party_name', 'family_name', 'date', 'month', 'calendar_week', 'issue_mono', 'issue_label', 'header', 'text'
]

# clean `df_de`
df_de = df_de[needed_vars]

-- *Remove outliers (based on word count in press releases)*

# add a word count column
df_de['n_words'] = df_de['text'].str.split().apply(len)

# define lower and upper cutoffs
lower_bound = df_de['n_words'].quantile(0.025)
upper_bound = df_de['n_words'].quantile(0.975)

# remove outliers (filter within range)
df_de = df_de[(df_de['n_words'] >= lower_bound) & (df_de['n_words'] <= upper_bound)].copy()

-- *Clean and save final dataset*

# add index for each unique press release
df_de['doc_id'] = df_de.reset_index().index + 1

# save as .csv
df_de.to_csv(os.path.join(ROOT_DIR, "out", "df_de.csv"), index=False)

# load `df_de` data
#df_de = pd.read_csv(os.path.join(ROOT_DIR, "out", "df_de.csv"))

# look at a press release
print(df_de['text'].iloc[0])

-- *Split press releases into sentences*

# define function to split press releases into sentences
def explode_sentences(df, text_col='text'):
    
    # apply sentence splitting
    df['sentences'] = df[text_col].apply(sent_tokenize)
    
    # each sentence gets its own row
    df_exploded = df.explode('sentences').reset_index(drop=True)
    
    # rename column
    df_exploded = df_exploded.rename(columns={'sentences': 'sentence'})
    
    return df_exploded

# run function
df_sentences = explode_sentences(df_de)

df_sentences.info()

# save as .csv
df_sentences.to_csv(os.path.join(ROOT_DIR, "out", "df_sentences.csv"), index=False)

# save as .pkl file
df_sentences.to_pickle(os.path.join(ROOT_DIR, "out", "df_sentences.pkl"))

-- *Extract only sentence text*

# extract sentence text
sentences = list(df_sentences['sentence'])

# save sentence data as .pkl file
with open(os.path.join(ROOT_DIR, "out", "sentences.pkl"), 'wb') as f:
    pickle.dump(sentences, f)
